Disable fixup_module_metadata when running Sphinx
